Extract chat confirmations into their own file
